Add STA thread for interacting with Visual Studio
This change updates the message filter to be within an STA thread. I couldn't repro it, but have had this issue in the past and it was fixed by running it within a dedicated thread that is marked for STA.

As part of this change, the TFMs are updated to be -windows since IIS usage is restricted to windows.

Fixes #5